+++ tests/models/test_documents.py
@@ -20,6 +20,7 @@ from caselawclient.models.documents import (
     CannotPublishUnpublishableDocument,
     Document,
     DocumentNotSafeForDeletion,
+    NonXMLDocumentError,
     UnparsableDate,
 )
 from caselawclient.models.judgments import Judgment
@@ -60,6 +61,37 @@ class TestDocument:
         document = Document("test/1234", mock_api_client)
 
         assert document.failed_to_parse is True
+
+    def test_xml_root_element_akomantoso(self, mock_api_client):
+        mock_api_client.get_judgment_xml_bytestring.return_value = "<akomaNtoso xmlns:uk='https://caselaw.nationalarchives.gov.uk/akn' xmlns='http://docs.oasis-open.org/legaldocml/ns/akn/3.0'>judgment</akomaNtoso>".encode(
+            "utf-8"
+        )
+
+        document = Document("test/1234", mock_api_client)
+
+        assert (
+            document.xml_root_element
+            == "{http://docs.oasis-open.org/legaldocml/ns/akn/3.0}akomaNtoso"
+        )
+
+    def test_xml_root_element_error(self, mock_api_client):
+        mock_api_client.get_judgment_xml_bytestring.return_value = (
+            "<error>parser.log contents</error>".encode("utf-8")
+        )
+
+        document = Document("test/1234", mock_api_client)
+
+        assert document.xml_root_element == "error"
+
+    def test_xml_root_element_malformed(self, mock_api_client):
+        mock_api_client.get_judgment_xml_bytestring.return_value = (
+            "<error>malformed xml".encode("utf-8")
+        )
+
+        document = Document("test/1234", mock_api_client)
+
+        with pytest.raises(NonXMLDocumentError):
+            document.xml_root_element
 
     def test_document_parsed(self, mock_api_client):
         mock_api_client.get_judgment_xml_bytestring.return_value = """